refactor the code base
