+++ .env.sample
@@ -1,0 +1,7 @@
+REACT_APP_API_KEY=<api key>
+REACT_APP_AUTH_DOMAIN=<auth domain>
+REACT_APP_DATABASE_URL=<database url>
+REACT_APP_PROJECT_ID=<project id>
+REACT_APP_STORAGE_BUCKET=<storage bucket>
+REACT_APP_MESSAGING_SENDER_ID=<messageing sender id>
+REACT_APP_APP_ID=<app id>
