+++ README.md
@@ -1,19 +1,33 @@
-# Mesh
+# Mesh Order App
 
 ## Description
 
-Mesh is an application that help you organize your orders. This was built with React.
+**`Mesh`** is an application that help you organize your orders. This was built with React.
+
+### Features
+* User Sign-in
+* All Order lists
+* View Single Order Details
+* Update Order
+* Logout User
+* Sign up page coming soon...
 
 ## Requirement
 
 Nodejs > 12.0.0
 
 ## Installation
+- Install Node
+- On your `terminal` or `CMD`
 - clone the repo
-    `git clone git@github.com:onwuvic/mesh.git` 
+    > git clone git@github.com:onwuvic/mesh.git
     or 
-    `git clone https://github.com/onwuvic/mesh.git`
+
+    > git clone https://github.com/onwuvic/mesh.git
+
 - cd into `mesh`
+    > cd mesh
+
 - run `yarn install`
 - run `yarn start`
 
@@ -28,7 +42,7 @@ You will also see any lint errors in the console.
 
 ### `yarn test`
 
-Launches the test runner in the interactive watch mode.\
+Currently no test for the frontend
 
 ### `yarn build`
 
@@ -38,5 +52,19 @@ It correctly bundles React in production mode and optimizes the build for the be
 The build is minified and the filenames include the hashes.\
 Your app is ready to be deployed!
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+## Technologies
+* [REACT](https://facebook.github.io/react/): REACT is a JavaScript framework developed by Facebook and it is used for developing web application. REACT is the 'VIEW' in the MVC architecture.
+* [Babel:](https://babeljs.io/): Babel is used to transpile es6 down to es5.
+* [Webpack:](https://webpack.github.io/docs/what-is-webpack.html): webpack is a module bundler.
+* [Axios:](https://www.npmjs.com/package/axios): Axios is a Promise based HTTP client for the browser and node.js
+
+## Language
+- Typescript
+
+## Author
+[Victor Onwuzor](https://github.com/onwuvic)
+
+## License & Copyright
+MIT © 
+
