+++ src/helpers/isLoggedIn.ts
@@ -1,0 +1,7 @@
+export const isLoggedIn = () => {
+    if (localStorage.getItem('token') !== null) {
+      return true;
+    }
+    return false;
+  };
+
